refactor(config-helper): moves normalizeSchema to own file; update normalized schema definition
