+++ packages/config-helper/src/index.ts
@@ -9,3 +9,4 @@ export * from './lib/params'
 export * from './lib/schema.error'
 export * from './lib/config.error'
 export * from './lib/loader'
+export * from './lib/schema-normalizer'
