+++ packages/config-helper/src/lib/normalized-schema.ts
@@ -1,22 +1,16 @@
-import { ConfigDefinition, ConfigValueTransformer } from './schema'
+import { ConfigDefinition, ConfigDefinitionCommon } from './schema'
 
-export interface NormalizedConfigDefinition<T> {
-	transformer: ConfigValueTransformer<T>
-	envVar: string
+export interface NormalizedConfigDefinition<T> extends Required<ConfigDefinitionCommon<T>> {
 	optional: boolean
-	defaultValue: T | null
-	trimValue: boolean | 'start' | 'end'
 }
 
 // eslint-disable-next-line @typescript-eslint/no-type-alias
 export type NormalizeSchema<T> = {
-	[P in keyof T]: T[P] extends ConfigDefinition<unknown>
-		? Omit<T[P], 'envVar'> & { envVar: string }
-		: NormalizeSchema<T[P]>
+	[P in keyof T]: T[P] extends ConfigDefinition<unknown> ? NormalizedConfigDefinition<T[P]> : NormalizeSchema<T[P]>
 }
 
 export function isNormalizedSchemaObject<T>(
 	x: NormalizeSchema<unknown> | NormalizedConfigDefinition<T>,
 ): x is NormalizedConfigDefinition<T> {
-	return 'transformer' in x
+	return 'envVar' in x && 'trimValue' in x
 }
